Pojo层调整电影对象实体类部分字段数据类型

+++ movie-backend/movie-pojo/src/main/java/com/samgoldsee/movie/entity/Movie.java
@@ -32,7 +32,7 @@ public class Movie implements Serializable {
     private Integer zoneId;
 
     // 浏览数
-    private Integer view;
+    private Long view;
 
     // 海报链接
     private String poster;
